Add fetch get stats.
