+++ massa-graph-2-exports/src/channels.rs
@@ -1,5 +1,3 @@
-use std::sync::mpsc::Receiver;
-
 use crossbeam_channel::Sender;
 use massa_execution_exports::ExecutionController;
 use massa_pool_exports::PoolController;
